Update patch set 4
Patch Set 4:

I'm fine with 1/2 being in 1 but I think we need to move the other details out separately so that they can be merged quickly.

There is more comment's I'd love to leave on this patch but don't want to distract the review with the minor stuff that can just land quickly.

Patch-set: 4