Add Jenkins Sandbox to index
For a doc that we use frequently it shouldn't take 2 clicks from the
landing page to get to. Move it to the index.

Add the word "Guide" to the doc so that it correctly appears on the
index page when linked.

Update the Jenkins page link to the Jenkins Sandbox to use a doc link.

Change-Id: I088c5ac175f36c3c14002987f66552f2bb8f604d
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>