Update patch set 5
Patch Set 5:

(2 comments)

a few minor typos and a general comment: as a newbie to JJB, I find this guide to be as confusing now as I did when I first read it last year when I wanted to test a new job I had created (I'm not a sysadmin and perhaps a tad brain dead when it comes to things like JJB). Suggestion: add a note to the top of the guide about minimum level of competence, something like "Users are expected to have a basic understanding of Jenkins and JJB." and add links to further info/tutorials. OR expand to give examples by scenario. Perhaps.... a job is broken in production and you want to push it to the sandbox to troubleshoot - use jjb-deploy. AND if you are creating a new job that you want to test out, install JJB and use the CLI. I personally would love to see much more detail, including a sample new job, screenshots of the terminal window, etc.

Patch-set: 5
Label: Verified=0