Add additional reading about good git messages
Since we like to give as many sources of good content as possible when
it comes to commit messages, another blog post was recently written and
brought brought to my attention. It's got additional links to some more
good content.

In this case it goes into detail on what a really good message looks
like, why it's good. For instance, it explains the reason behind the
change and all the work done to implement it along with giving learning
examples.

Change-Id: Id4fef32e15beededb6c13197d8e85922c5e808a9
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>