Update analytics dashboard links
With the rebrand to LFX the URLs have all been changed for
lfanalytics.io. While there is a redirect in place it breaks for many of
the URLs because of %2F in the URI and a bug at DNSimple.

Instead of relying on the redirect, we should have the dashboards linked
properly

Additionally update Acumos Sonar link to only be SonarCloud

Change-Id: Ia122a52d9d7df23ba0570952e68af744bcf4a533
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>