Docs: Add Conventional Commit specification links
While creating a Current Best Practices presentation for some Computer
Science students I stumbled upon references to Semantic Commit Messages.

Upon further reading on the subject I believe that this is a new
practice that we should be recommending to communities. There seems to
be a rabbit hole of references on it, so I'm only recommending the two
that seem to give the best results. In particular the Gist has continued
discussion even after the initial posting on May 15, 2017.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>
Change-Id: I3669af717222172a7288052ff838c71e35c6524c