CI!: Enforce gitlint in tox
gitlint current configuration allows to check that commit messages abide
by the conventional commit format as recently recommended by best
practices.

This is currently configured in the pre-commit job but it is not
performed by the gate yet

A dedicated tox profile is proposed here to automate this task in
the gate and to detect non-compliant changes.

Current best practices references such as Chris Beams' blog recommends
to start subject line with a capital letter.
https://chris.beams.io/posts/git-commit/
With conventional commits, this is classically no more true by default.
Subject lines classicaly starts with a type such as fix: or feat:
that are in lower-cases as noticed by the Gnome project
https://wiki.gnome.org/Git/CommitMessages

For this reason, in Releng, the types slightly differ from the common
practice and all starts with a capital letter.
They are described in the file .gitlint.
cf If45c7a256f7d769eb092d867217468b9d6c53384
and are currently:
Fix, Feat, Chore, Docs, Style, Refactor, Perf, Test, Revert, CI, Build

Signed-off-by: Guillaume Lambert <guillaume.lambert@orange.com>
Change-Id: Ie43773c374e35343edd8fa05600e95644eea671a