CI: Properly run gitlint in CI
When running pre-commit in a CI system the COMMIT_EDITMSG does not
(normally) get created as that is an artifact of editing the commit
message. If the file doesn't exist then gitlint will skip which makes it
possible for pre-commit checks that should fail, to pass.

Since we want tox to run in a consistent manner both locally and in CI
we need to play around a little with how we are checking the commit
message.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>
Change-Id: Ife7b4f04be2bad582d2f406c9c72bf772b097b21